Restructure project file structure
